refactor: keep createPcg32 in pcg32.ts for variant symmetry
Moves createPcg32 (and the MASK_64 constant + resolveStreamScheme
helper it uses) back into pcg32.ts so each variant module owns its
own create<Variant> factory, mirroring mulberry32.ts. index.ts now
just re-exports createPcg32 and defines the createPcg legacy alias.
